+++ projects/ng-wizi-bulma/src/lib/shared/index.ts
@@ -1,3 +1,4 @@
 export * from './components/index';
 export * from './directives/index';
+export * from './services/index';
 export * from './common-module';