+++ projects/ng-wizi-bulma/src/lib/shared/services/index.ts
@@ -1,0 +1,2 @@
+export * from './filter-routing/model';
+export * from './filter-routing/filter-routing-group.service';